+++ docs/conf.py
@@ -30,6 +30,7 @@ copyright = f'{datetime.now().year}, theOehrly'
 # ones.
 extensions = [
     'sphinx.ext.autodoc',
+    'sphinx.ext.autosummary',
     'sphinx.ext.todo',
     'sphinx.ext.viewcode',
     'sphinx.ext.napoleon',